Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main() {
-    float a,b;
+    double a,b;
     printf(" ");
-    scanf("%f%f",&a,&b);
-    printf("You entered: %.4f", a,b);
+    scanf("%lf%lf",&a,&b);
+    printf("You entered: %.4lf", a,b);
     return 0;
 }
